% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{bulma_button}
\alias{bulma_button}
\alias{bulma_buttons}
\alias{bulma_delete}
\title{Create a bulma Button}
\usage{
bulma_button(
  label = NULL,
  icon = NULL,
  color = NULL,
  light = FALSE,
  size = NULL,
  icon_size = "normal",
  icon_class = NULL,
  icon_lib = "font-awesome",
  icon_pos = c("start", "end"),
  full_width = FALSE,
  disabled = FALSE,
  style = NULL,
  state = NULL,
  container = htmltools::tags$button,
  ...
)

bulma_buttons(
  ...,
  alignment = NULL,
  addons = FALSE,
  size = NULL,
  container = htmltools::div
)

bulma_delete(size = NULL)
}
\arguments{
\item{label}{[\code{character(1)}: \sQuote{NULL}]\cr
The label of the button.}

\item{icon}{[\code{character(1)}: \sQuote{NULL}]\cr
The name of the icon.}

\item{color}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma color name.}

\item{light}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the \code{is-light} modifier is used, resulting in a lighter color.}

\item{size}{[\code{character(1)}: \sQuote{NULL}]\cr
The size of the button or the button list. Valid sizes are
\code{small}, \code{normal}, \code{medium}, \code{large}.}

\item{icon_size, icon_pos, icon_class, icon_lib}{[\code{character(1)}: \sQuote{normal},
\sQuote{start} and \sQuote{NULL} respectively]\cr
Arguments passed down \code{bulma_icon} controlling the appearance of the icon.}

\item{full_width}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the button spans the full width.}

\item{disabled}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE}, the button will be disabled.}

\item{style}{[\code{character(n)}: \sQuote{NULL}]\cr
Style modifiers to be applied to the button. Several modifiers can be combined.
Valid styles are \code{outlined}, \code{inverted} and \code{rounded}.}

\item{state}{[\code{character(1)}: \sQuote{NULL}]\cr
State modifier to be applied to the button. Valid states are \code{hovered},
\code{focused}, \code{active}, \code{selected}, \code{loading} and \code{static}.}

\item{container}{[\code{function}]\cr
The container to be used for the \code{button} or the button list.
For the \code{button} it Should be a function creating either
a \verb{<button>}, an \verb{<a>} or an \verb{<input type = "submit|reset">} HTML tag.}

\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Further arguments passed down to \code{container}.
\subsection{Note}{

This function creates the HTML to render a bulma button. It does \strong{not}, however, add
any reactivity.
}}

\item{alignment}{[\code{character(1)}: \sQuote{NULL}]\cr
The alignment to be used for the button list. Valid alignments are \code{centered} or
\code{right}.}

\item{addons}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the buttons in the button list are attached together.}
}
\value{
A bulma \code{button}, a \code{delete} button or a buttons list.
}
\description{
Create a bulma Button
}
\examples{
## Only run examples in interactive R sessions
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
   ui <- bulma_page(
     bulma_block(
       bulma_title("Simple Button"),
       bulma_button("Button")
     ),
     bulma_block(
       bulma_title("Different Containers"),
       bulma_button("Anchor", container = a),
       bulma_button(container = function(...) tags$input(..., type = "submit",
                                                         value = "Submit Input")),
       bulma_button(container = function(...) tags$input(..., type = "reset",
                                                         value = "Reset Input"))
     ),
     bulma_block(
       bulma_title("Different Colors"),
       bulma_button("White", color = "white"),
       bulma_button("Light", color = "light"),
       bulma_button("Dark", color = "dark"),
       bulma_button("Black", color = "black"),
       bulma_button("Text", color = "text"),
       bulma_button("Ghost", color = "ghost")
     ),
     bulma_block(
       bulma_title("Button Groups"),
       bulma_buttons(
         bulma_button("Primary", color = "primary"),
         bulma_button("Link", color = "link")
       ),
       bulma_buttons(
         bulma_button("Info", color = "info"),
         bulma_button("Success", color = "success"),
         bulma_button("Warning", color = "warning"),
         bulma_button("Danger", color = "danger")
       )
     ),
     bulma_block(
       bulma_title("Light Colors"),
       bulma_buttons(
         bulma_button("Primary", color = "primary", light = TRUE),
         bulma_button("Link", color = "link", light = TRUE)
       ),
       bulma_buttons(
         bulma_button("Info", color = "info", light = TRUE),
         bulma_button("Success", color = "success", light = TRUE),
         bulma_button("Warning", color = "warning", light = TRUE),
         bulma_button("Danger", color = "danger", light = TRUE)
       )
     ),
     bulma_block(
       bulma_title("Different Sizes"),
       bulma_button("Small", size = "small"),
       bulma_button("Normal [Default]", size = "normal"),
       bulma_button("Medium", size = "medium"),
       bulma_button("Large", size = "large"),
       bulma_buttons(
         size = "medium",
         bulma_button("All"),
         bulma_button("Medium"),
         bulma_button("Size")
       ),
       bulma_buttons(
         size = "small",
         bulma_button("Small"),
         bulma_button("Small"),
         bulma_button("Small"),
         bulma_button("Normal", size = "normal"),
         bulma_button("Small")
       )
     ),
     bulma_block(
       bulma_title("Fullwidth"),
       bulma_button("Small", size = "small", full_width = TRUE),
       bulma_button("Normal", full_width = TRUE),
       bulma_button("Medium", size = "medium", full_width = TRUE),
       bulma_button("Large", size = "large", full_width = TRUE)
     ),
     bulma_block(
       bulma_title("Different Styles"),
       h2("Outlined", class = "subtitle"),
       bulma_button("Outlined", style = "outlined"),
       bulma_button("Outlined", color = "primary", style = "outlined"),
       bulma_button("Outlined", color = "link", style = "outlined"),
       bulma_button("Outlined", color = "info", style = "outlined"),
       bulma_button("Outlined", color = "danger", style = "outlined"),
       bulma_button("Outlined", color = "warning", style = "outlined"),
       bulma_button("Outlined", color = "success", style = "outlined")
     ),
     bulma_block(
       h2("Inverted", class = "subtitle"),
       bulma_button("Inverted", style = "inverted"),
       bulma_button("Inverted", color = "primary", style = "inverted"),
       bulma_button("Inverted", color = "link", style = "inverted"),
       bulma_button("Inverted", color = "info", style = "inverted"),
       bulma_button("Inverted", color = "danger", style = "inverted"),
       bulma_button("Inverted", color = "warning", style = "inverted"),
       bulma_button("Inverted", color = "success", style = "inverted")
     ),
     bulma_block(
       h2("Rounded", class = "subtitle"),
       bulma_button("Rounded", style = "rounded"),
       bulma_button("Rounded", color = "primary", style = "rounded"),
       bulma_button("Rounded", color = "link", style = "rounded"),
       bulma_button("Rounded", color = "info", style = "rounded"),
       bulma_button("Rounded", color = "danger", style = "rounded"),
       bulma_button("Rounded", color = "warning", style = "rounded"),
       bulma_button("Rounded", color = "success", style = "rounded")
     ),
     bulma_block(
       bulma_title("Different States"),
       bulma_button("Hovered", state = "hovered"),
       bulma_button("Focused", color = "primary", state = "focused"),
       bulma_button("Active", color = "link", state = "active"),
       bulma_button("Selected", color = "info", state = "selected"),
       bulma_button("Loading", color = "danger", state = "loading"),
       bulma_button("Static", color = "warning", state = "static"),
       bulma_button("Disabled", color = "success", disabled = TRUE)
     ),
     bulma_block(
       bulma_title("Icons"),
       bulma_buttons(
         bulma_button(icon = "bold", icon_class = "fas"),
         bulma_button(icon = "italic", icon_class = "fas"),
         bulma_button(icon = "underline", icon_class = "fas"),
         container = p
       ),
       bulma_buttons(
         bulma_button("GitHub", "github", icon_class = "fab"),
         bulma_button("@jgthms", "twitter", color = "primary", icon_class = "fab"),
         bulma_button("Save", "check", color = "success", icon_class = "fab",
                      icon_size = "small"),
         bulma_button("Delete", "times", color = "danger", style = "outlined",
                      class = "fab", icon_size = "small", icon_pos = "end"),
         container = p
       ),
       bulma_buttons(
         bulma_button("GitHub", "github", icon_class = "fab", size = "small",
                      icon_size = "small"),
         bulma_button("GitHub", "github", icon_class = "fab"),
         bulma_button("GitHub", "github", icon_class = "fab", size = "medium"),
         bulma_button("GitHub", "github", icon_class = "fab", size = "large",
                      icon_size = "medium"),
         container = p
       ),
       bulma_buttons(
         bulma_button(icon = "heading", icon_class = "fas", size = "small",
                      icon_size = "small"),
         container = p
       ),
       bulma_buttons(
         bulma_button(icon = "heading", icon_class = "fas", icon_size = "small"),
         bulma_button(icon = "heading", icon_class = "fas fa-lg"),
         container = p
       ),
       bulma_buttons(
         bulma_button(icon = "heading", icon_class = "fas", size = "medium",
                      icon_size = "small"),
         bulma_button(icon = "heading", icon_class = "fas fa-lg", size = "medium"),
         bulma_button(icon = "heading", icon_class = "fas fa-2x", size = "medium",
                      icon_size = "medium"),
         container = p
       ),
       bulma_buttons(
         bulma_button(icon = "heading", icon_class = "fas", size = "large",
                      icon_size = "small"),
         bulma_button(icon = "heading", icon_class = "fas fa-lg", size = "large"),
         bulma_button(icon = "heading", icon_class = "fas fa-2x", size = "large",
                      icon_size = "medium"),
         container = p
       )
     ),
     bulma_block(
       bulma_title("List of buttons"),
       bulma_buttons(
         bulma_button("Yes"),
         bulma_button("Maybe"),
         bulma_button("No"),
         addons = TRUE
       ),
       bulma_buttons(
         bulma_button("Yes"),
         bulma_button("Maybe"),
         bulma_button("No"),
         addons = TRUE,
         alignment = "centered"
       ),
       bulma_buttons(
         bulma_button("Yes"),
         bulma_button("Maybe"),
         bulma_button("No"),
         addons = TRUE,
         alignment = "right"
       ),
       bulma_buttons(
         bulma_button("Yes", color = "success", state = "selected"),
         bulma_button("Maybe"),
         bulma_button("No"),
         addons = TRUE
       ),
       bulma_buttons(
         bulma_button("Yes"),
         bulma_button("Maybe", color = "link", state = "selected"),
         bulma_button("No"),
         addons = TRUE
       ),
       bulma_buttons(
         bulma_button("Yes"),
         bulma_button("Maybe"),
         bulma_button("No", color = "danger", state = "selected"),
         addons = TRUE
       )
     ),
     bulma_block(
       bulma_title("Delete Button"),
       bulma_delete(size = "small"),
       bulma_delete(),
       bulma_delete(size = "medium"),
       bulma_delete(size = "large")
     )
   )

   server <- function(input, output) {
   }

   shinyApp(ui, server)
}
}
\seealso{
\link{bulma_icon}, \href{https://bulma.io/documentation/elements/button/}{Bulma Button}
}
