% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{bulma_columns}
\alias{bulma_columns}
\alias{bulma_column}
\title{Create a Columns Layout}
\usage{
bulma_columns(...)

bulma_column(..., width = NULL, offset = NULL, media_breakpoint = NULL)
}
\arguments{
\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Elements to include within the column.}

\item{width}{[\code{character(1)} or \code{integer(1)}: \sQuote{NULL}]\cr
The width to be used (see \link{validate_bulma_column_size} for valid values).
If \code{NULL}, all available space is used. If several columns use \code{NULL}, the
space is distributed among those columns.}

\item{offset}{[\code{character(1)} or \code{integer(1)}: \sQuote{NULL}]\cr
The offset to be used (see \link{validate_bulma_column_offset} for valid values).}

\item{media_breakpoint}{[\code{character(1)}: \sQuote{NULL}]\cr}
}
\value{
A Bulma column layout.
}
\description{
Create a Columns Layout
}
\examples{

## Only run examples in interactive R sessions
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {

ui <- bulma_page(
        bulma_columns(
           bulma_column(tags$h1("Column 1"), width = 3, offset = 2,
                        class = "has-background-primary"),
           bulma_column(tags$h1("Column 2"), width = 4, offset = 1,
                        class = "has-background-link")
        )
      )
server <- function(input, output) {
}

shiny::shinyApp(ui, server)
}
}
\seealso{
\link{validate_bulma_column_size},\link{validate_bulma_column_offset},
\link{validate_bulma_media_breakpoint},
\href{https://bulma.io/documentation/columns/basics/}{Bulma Columns}
}
