% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{bulma_container}
\alias{bulma_container}
\title{Create a bulma Container}
\usage{
bulma_container(
  ...,
  fluid = FALSE,
  fullwidth_breakpoint = NULL,
  narrow_down_to = NULL
)
}
\arguments{
\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Elements to include within the column.}

\item{fluid}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE}, the \code{is-fluid} class is added to teh container, resulting in
a container without a maximum width.}

\item{fullwidth_breakpoint}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma media breakpoint until which the container is shown in full width.
Must be either \code{widescreen} or \code{fullhd}.}

\item{narrow_down_to}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma breakpoint to whose maximum size the container is narrowed down
to. Must be either \code{desktop} or \code{widescreen}.}
}
\value{
A bulma \code{container}.
}
\description{
Create a bulma Container
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  ui <- bulma_page(
    bulma_block(
      bulma_title("Simple Container"),
      bulma_container(
        bulma_notification(
          tagList(
            "This is container is ",
            strong("centered"),
            " on desktop and larger viewports"),
          color = "primary",
          add_delete = FALSE)
      )
    ),
    bulma_block(
      bulma_title("Widescreen or FullHD only"),
      bulma_block(
        bulma_container(
          bulma_notification(
            tagList(
              "This container is ",
              strong("fullwidth"),
              " ",
              tags$em("until"),
              " the ",
              tags$code("$widescreen"),
              " breakpoint."
            ),
            color = "primary",
            add_delete = FALSE
          ),
          fullwidth_breakpoint = "widescreen"
        )
      ),
      bulma_block(
        bulma_container(
          bulma_notification(
            tagList(
              "This container is ",
              strong("fullwidth"),
              " ",
              tags$em("until"),
              " the ",
              tags$code("$fullhd"),
              " breakpoint."
            ),
            color = "primary",
            add_delete = FALSE
          ),
          fullwidth_breakpoint = "fullhd"
        )
      )
    ),
    bulma_block(
      bulma_title("Desktop and Widescreen maximum widths"),
      bulma_block(
        bulma_container(
          bulma_notification(
            tagList(
              "This container has a ",
              tags$code("max-width"),
              " of ",
              tags$code("$desktop - $container-offset"),
              " on widescreen and fullhd."
            ),
            color = "primary",
            add_delete = FALSE
          ),
          narrow_down_to = "desktop"
        )
      ),
      bulma_block(
        bulma_container(
          bulma_notification(
            tagList(
              "This container has a ",
              tags$code("max-width"),
              " of ",
              tags$code("$widescreen - $container-offset"),
              " on fullhd."
            ),
            color = "primary",
            add_delete = FALSE
          ),
          narrow_down_to = "widescreen"
        )
      )
    ),
    bulma_block(
      bulma_title("Fluid container"),
      bulma_container(
        bulma_notification(
          tagList(
            "This is container is ",
            strong("fluid"),
            "; it will have a 32px gap on either side, on any viewport sie."),
          color = "primary",
          add_delete = FALSE),
        fluid = TRUE
      )
    )
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/layout/container/}{Bulma Container}
}
