% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{bulma_header}
\alias{bulma_header}
\alias{bulma_title}
\alias{bulma_subtitle}
\title{Create a bulma Title or Subtitle}
\usage{
bulma_header(
  ...,
  container = if (type == "title") htmltools::h1 else htmltools::h2,
  size = NULL,
  type = c("title", "subtitle"),
  maintain_space = FALSE
)

bulma_title(
  ...,
  container = htmltools::h1,
  size = NULL,
  maintain_space = FALSE
)

bulma_subtitle(..., container = htmltools::h2, size = NULL)
}
\arguments{
\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Elements to include within the container.}

\item{container}{[\code{function}\: \sQuote{h1} or \sQuote{h2}]\cr
The container for the title. Typically one of \code{htmltools::h[1-6]}.}

\item{size}{[\code{numeric(1)}: \sQuote{NULL}]\cr
The size for the title. A valid size is an integer between 1 and 6.}

\item{type}{[\code{character(1)}: \sQuote{NULL}]\cr
The type of the heading, must be either \sQuote{title} or \sQuote{subtitle}.}

\item{maintain_space}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the sapce between a title and a subtitle is maintained.
Otherwise they move closer together.}
}
\value{
A bulma title or subtitle.
}
\description{
\code{bulma_header} creates a title or a subtitle for bulma.
}
\details{
\code{bulma_title} and \code{bulma_subtitle} are convenience wrappers which set \code{type}
accordingly.
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
   ui <- bulma_page(
     bulma_block(
       bulma_title("Basic Usage",
                   size = 1,
                   class = make_class("text-primary", prefix = "has")),
       bulma_title("A Title"),
       bulma_subtitle("A Subtitle")
     ),
     bulma_block(
       bulma_title("Different Sizes",
                   size = 1,
                   class = make_class("text-primary", prefix = "has")),
       Map(function(size) {
         text <- paste("Title", size)
         if (size == 3) {
           text <- paste(text, "(default size)")
         }
         bulma_title(text, size = size)
       }, 1:6),
       Map(function(size) {
         text <- paste("Subtitle", size)
         if (size == 5) {
           text <- paste(text, "(default size)")
         }
         bulma_title(text, size = size)
       }, 1:6)
     ),
     bulma_block(
       bulma_title("Combine Title and Subtitle",
                   size = 1,
                   class = make_class("text-primary", prefix = "has")),
       Map(function(size) {
         htmltools::tagList(
           bulma_title(paste("Title", size), size = size),
           bulma_subtitle(paste("Subtitle", size + 2), size = size + 2)
         )
       }, 1:3)
     ),
     bulma_block(
       bulma_title("Maintain Space Between Title and Subtitle",
                   size = 1,
                   class = make_class("text-primary", prefix = "has")),
       Map(function(size) {
         htmltools::tagList(
           bulma_title(paste("Title", size), size = size, maintain_space = TRUE),
           bulma_subtitle(paste("Subtitle", size + 2), size = size + 2)
         )
       }, 1:3)
     )
   )

   server <- function(input, output) {
   }

   shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/elements/title/}{Bulma Titles}
}
