% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hero.R
\name{bulma_hero}
\alias{bulma_hero}
\title{Title}
\usage{
bulma_hero(
  title = NULL,
  subtitle = NULL,
  body = NULL,
  header = NULL,
  footer = NULL,
  color = NULL,
  size = NULL,
  with_navbar = FALSE,
  container = htmltools::tags$section,
  ...
)
}
\arguments{
\item{title, subtitle, body}{[\code{shiny.tag} or \code{listish}; \sQuote{NULL}]\cr
The title and the subtitle of the bulma hero banner. Either one specifies
\code{title} and \code{subtitle} to display a title and subtitle on the hero banner,
or one uses \code{body} to freely style the content. If, for instance, you want to
use a different size for the title, you can also use
\code{body = bulma_title("Title", size = 1)}.}

\item{header, footer}{[\code{shiny.tag} or \code{listish}; \sQuote{NULL}]\cr
The header and footer of the bulma hero banner. Can be used wit
\code{title}/\code{subtitle}  and \code{body}.}

\item{color}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma color name.}

\item{size}{[\code{character(1)}: \sQuote{NULL}]\cr
The size of the hero banner. Valid sizes are
\code{small}, \code{medium}, \code{large}, \code{halfheight} and \code{fullheight}.}

\item{with_navbar}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE}, the hero banner will be made smaller to allow space for
the navabr.}

\item{container}{[\code{function}]\cr
The container to be used for the bulma hero banner.}

\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Further arguments passed down to \code{container}.}
}
\value{
A bulma hero banner.
}
\description{
Title
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  capitalize <- function(x) paste0(toupper(substr(x, 1, 1)),
                                   substr(x, 2, nchar(x)))

  ui <- bulma_page(
    bulma_block(
      bulma_title("Colors"),
      lapply(bulma_config[bulma_config$is_color_map_key, "variable"],
             function(color) bulma_block(
               bulma_hero(paste(capitalize(color), "hero"),
                          paste(capitalize(color), "subtitle"),
                          color = color)
             )
      )
    ),
    bulma_block(
      bulma_title("Sizes"),
      Map(function(size, color) bulma_block(
        bulma_hero(
          paste(capitalize(size), "hero"),
          paste(capitalize(size), "subtitle"),
          size = size,
          color = color
        )
      ),
      c("small", "medium", "large", "halfheight", "fullheight"),
      c("primary", "link", "info", "success", "danger"))
    )
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)

  navbar <- tags$nav(
    class = "navbar",
    bulma_container(
      div(class = "navbar-menu",
          div(class = "navbar-start",
              tags$a(class = "navbar-item", "Home"),
              tags$a(class = "navbar-item", "Documentation")
          ),
          div(class = "navbar-end",
              div(class = "navbar-item",
                  bulma_buttons(
                    bulma_button("Github", color = "dark", container = tags$a),
                    bulma_button("Download", color = "link", container = tags$a)
                  )
              )

          )
      )
    )
  )

  ui <- bulma_page(
    navbar,
    bulma_hero(
      "Fullheight hero with navbar",
      color = "link",
      with_navbar = TRUE
    )
  )

  shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/layout/hero/}{Bulma Hero Banner}
}
