% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{bulma_icon}
\alias{bulma_icon}
\title{Create bulma Icon Container}
\usage{
bulma_icon(
  name,
  text = NULL,
  color = NULL,
  size = NULL,
  container = htmltools::span,
  class = NULL,
  lib = "font-awesome",
  ...
)
}
\arguments{
\item{name}{[\code{character(n)}]\cr
Name of the icons.}

\item{text}{[\code{character(n)}; \sQuote{NULL}]\cr
Text to be displayed next to the icon. Must be of the same length as \code{name} or
\sQuote{NULL}, in which case a single icon without any text is created. If you
want to skip some text, provide \code{NA}.}

\item{color}{[\code{character(1)}; \sQuote{NULL}]\cr
A valid bulma color name.}

\item{size}{[\code{character(1)}; \sQuote{NULL}]\cr
The size of the icon \emph{container}. The size of the icon itself is controlled via
arguments passed to \link[shiny:icon]{shiny::icon} and depends on the icon library used.}

\item{container}{[\code{function}]\cr
Either \code{div} or \code{span}. The parent element of the icon. Only relevant, if we
want to display text next to the icon.}

\item{class, lib, ...}{[\code{character(1)} and \verb{html tags} or \verb{html attributes}]\cr
These arguments are passed down to \link[shiny:icon]{shiny::icon}.}
}
\value{
A bulma icon container optionally ammended by some text.
}
\description{
Create bulma Icon Container
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  ui <- bulma_page(
    tags$head(tags$style(HTML("#sizes > span {vertical-align: middle;}"))),
    bulma_block(
      bulma_title("Icon with Text"),
      bulma_icon("home", "Home")),
    bulma_block(
      bulma_title("Several Icons with Text"),
      bulma_icon(c("train", "arrow-right", "arrow-right",
                   "arrow-right", "flag-checkered"),
                 c("Paris", "Budapest", "Bucharest", "Istanbul", NA))),
    bulma_block(
      bulma_title("Inline Icon"),
      p("An invitation to", bulma_icon("utensils", "dinner"), "was soon afterwards ",
        "dispatched; and already had Mrs. Bennet planned the courses that were to do",
        "credit to her housekeeping, when an answer arrived which deferred it all.",
        "Mr. Bingley was obliged to be in", bulma_icon("city", "town"),
        "the following day, and, consequently, unable to accept the honour of their",
        bulma_icon("envelope-open-text", "invitation"), ", etc.")
    ),
    bulma_block(
      bulma_title("Icon With Text, Color in a Different Container"),
      bulma_icon("info-circle", "Information", color = "link", container = div),
      bulma_block("Your package will be delivered on", strong("Tuesday, 08:00")),
      bulma_icon("check-square", "Success", color = "success", container = div),
      bulma_block("Your image has been successfully uploaded.")
    ),
    bulma_block(
      id = "sizes",
      bulma_title("Icons with Sizing"),
      tagAppendAttributes(bulma_icon("home", size = "medium", class = "fas"),
                          class = "has-background-warning"),
      tagAppendAttributes(bulma_icon("home", size = "medium", class = "fas fa-lg"),
                          class = "has-background-warning"),
      tagAppendAttributes(bulma_icon("home", size = "medium", class = "fas fa-2x"),
                          class = "has-background-warning")
    )
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
\seealso{
\link{validate_bulma_color},
\href{https://bulma.io/documentation/elements/icon/}{Bulma Icons}
}
