% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{bulma_image}
\alias{bulma_image}
\title{Create a Responsive Figure}
\usage{
bulma_image(
  src,
  alt = NULL,
  ...,
  fixed = NULL,
  ratio = NULL,
  rounded = FALSE,
  use_fullwidth = FALSE,
  container = htmltools::tags$figure
)
}
\arguments{
\item{src}{[\code{character(1)}]\cr
Path to the image to be displayed.}

\item{alt}{[\code{character(1)}: \sQuote{NULL}]\cr
Alternative text description for the image.}

\item{...}{[\verb{named arguments}]\cr
Further attributes to be passed to the \verb{<img>} tag.}

\item{fixed}{[\code{integer(1)}: \sQuote{NULL}]\cr
A valid size for a \emph{fixed sized} bulma image or \sQuote{NULL}. Valid sizes are
16, 24, 32, 48, 64, 96 and 128.}

\item{ratio}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid ratio for a \emph{ratio sized} bulma image or \sQuote{NULL}. Valid ratios are
\code{square}, \verb{1by1}, \verb{5by4}, \verb{4by3}, \verb{3by2}, \verb{5by3}, \verb{16by9}, \verb{2by1}, \verb{3by1},
\verb{4by5}, \verb{3by4}, \verb{2by3}, \verb{3by5}, \verb{9by16}, \verb{1by2}, \verb{1by3}.}

\item{rounded}{[\code{logical(1)}: \sQuote{FALSE}]\cr
Should the image be rounded?}

\item{use_fullwidth}{[\code{logical(1)}: \sQuote{FALSE}]\cr
By default, the image container takes up the full width while maintaining the
given ratio. If it does not, you can force it by setting \code{use_fullwidth} to
\sQuote{TRUE}.}

\item{container}{[\code{function}]\cr
The container to be used for the bulma image.}
}
\value{
A bulma \code{figure} tag containing the given \verb{<img>}.
}
\description{
The bulma \code{image} class creates a container with either a fixed (squared) size or a
defined ratio. The former is typically used for avatars, while the latter can display
arbitrary sized images (with the only requirement that the parent has a defined
\code{width}). Either way, the container is precisely sized, which prevents the design to
break or wobble in case an image needs longer to load or does not load at all.
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  ## some blind text
  li <- paste("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam",
              "nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam",
              "erat, sed diam voluptua. At vero eos et accusam et justo duo dolores",
              "et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est",
              "Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur",
              "sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et",
              "dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam",
              "et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea",
              "takimata sanctus est Lorem ipsum dolor sit amet.")

  ## show the effect of bulma_image:
  ## the right column does not change in size while the left does as soon as the
  ## image is loaded

  ## delayed placeholder image
  src <- "http://www.deelay.me/2000/http://placehold.it/800x600"
  ui <- bulma_page(div(class = "container",
                       bulma_columns(bulma_column(tags$figure(img(src = src)), li),
                                     bulma_column(bulma_image(src, ratio = "4by3"),
                                                  li))))
  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/elements/image/}{Bulma Image}
}
