% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bulma_page}
\alias{bulma_page}
\title{Create a Bulma Page}
\usage{
bulma_page(..., title = NULL, theme = NULL, lang = NULL)
}
\arguments{
\item{...}{Elements to include within the page}

\item{title}{The browser window title (defaults to the host URL of the page).}

\item{theme}{If \code{NULL} the original bulma theme is used. Otherwise, it must be a
valid bulma theme name from
\href{https://jenil.github.io/bulmaswatch/}{bulmaswatch}.}

\item{lang}{ISO 639-1 language code for the HTML page, such as \dQuote{en} or
\dQuote{de}. This will be used as the lang in the \code{<html>} tag, as in
\code{<html lang="en">}. The default (\code{NULL}) results in an empty
string.}
}
\value{
A UI definition that can be passed to \code{\link[shiny]{shinyApp}}.
}
\description{
This function creates a bulma page and allows themes from bulmsawatch to be used.
}
\examples{

## Only run examples in interactive R sessions
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {

ui <- bulma_page(
        tags$section(class = "section",
                     div(class = "container",
                         bulma_title("Hello World!"),
                         bulma_subtitle("My first website with",
                                tags$strong("Bulma", .noWS = "after"), "!")))
      )
server <- function(input, output) {
}

shiny::shinyApp(ui, server)
}
}
\seealso{
\code{\link{get_bulma_themes}} for getting a list of all installed themes,
\url{https://bulma.io}
}
