% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{bulma_progressbar}
\alias{bulma_progressbar}
\title{Create a Bulma Progress Bar}
\usage{
bulma_progressbar(value = NULL, max = 100, color = NULL, size = NULL)
}
\arguments{
\item{value}{[\code{numeric(1)}: \sQuote{NULL}]\cr
The value of the progress bar. If \sQuote{NULL} the progress bar shows an
indeterminate progress.}

\item{max}{[\code{numeric(1)}: \sQuote{100}]\cr
The maximum value of the progress bar.}

\item{color}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma color name.}

\item{size}{[\code{character(1)}: \sQuote{NULL}]\cr
The size of the progress bar. Valid sizes are
\code{small}, \code{normal}, \code{medium}, \code{large}.}
}
\value{
A bulma progress bar.
}
\description{
Create a Bulma Progress Bar
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
   cols <- bulma_config[bulma_config$is_color_map_key, "variable"]
   perc <- 10 * seq_along(cols)
   ui <- bulma_page(
     bulma_block(
       bulma_title("Basic Progressbar"),
       bulma_progressbar(15)
     ),
     bulma_block(
       bulma_title("Colored Progressbars"),
       Map(function(val, col) bulma_progressbar(val, color = col),
           perc[-c(5, 10)], cols[-c(5, 10)])
     ),
     bulma_block(
       bulma_title("Different Sizes"),
       Map(function(val, size) bulma_progressbar(val, size = size),
           15 * 1:4, c("small", "normal", "medium", "large"))
     ),
     bulma_block(
       bulma_title("Indeterminate Progressbar"),
       Map(function(col) bulma_progressbar(color = col), cols)
     )
   )

   server <- function(input, output) {
   }

   shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/elements/progress/}{Bulma Progressbar}
}
