% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{bulma_tag}
\alias{bulma_tag}
\alias{bulma_tags}
\title{Create a bulma Tag}
\usage{
bulma_tag(
  label = NULL,
  color = NULL,
  light = FALSE,
  size = NULL,
  style = NULL,
  add_delete = FALSE,
  delete_size = "small",
  container = htmltools::span,
  ...
)

bulma_tags(..., addons = FALSE, size = NULL, container = htmltools::div)
}
\arguments{
\item{label}{[\code{character(1)}: \sQuote{NULL}]\cr
The label of the tag.}

\item{color}{[\code{character(1)}: \sQuote{NULL}]\cr
A valid bulma color name.}

\item{light}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the \code{is-light} modifier is used, resulting in a lighter color.}

\item{size}{[\code{character(1)}: \sQuote{NULL}]\cr
The size of the tag or the tag list. Valid sizes are
\code{normal}, \code{medium}, \code{large}.}

\item{style}{[\code{character(1)}: \sQuote{NULL}]\cr
Can be either \sQuote{NULL}, \sQuote{delete} or \sQuote{rounded} which
potentially transforms the tag to a \dQuote{delete} tag or a rounded tag.}

\item{add_delete}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} a delete button is added to the right of the tag.}

\item{delete_size}{[\code{character(1)}: \sQuote{small}]\cr
The size of the additional delete button.}

\item{container}{[\code{function}]\cr
The container to be used for the \code{tag} or the tag list.
For the \code{tag} it should be a function creating either
a \verb{<span>} or an \verb{<a>} HTML tag.}

\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Further arguments passed down to \code{container}.}

\item{addons}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE} the tags in the tag list are attached together.}
}
\value{
A bulma \code{tag} or a tag list.
}
\description{
Create a bulma Tag
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  numbers <- c("One", "Two", "Three", "Four", "Five", "Six", "Seven", "Eight", "Nine",
               "Ten", "Eleven", "Twelve", "Thirteen", "Fourteen", "Fifteen",
               "Sixteen", "Seventeen", "Eighteen", "Nineteen", "Twenty")
  ui <- bulma_page(
    bulma_block(
      bulma_title("Simple Tag"),
      bulma_button("Tag label")
    ),
    bulma_block(
      bulma_title("Different Colors"),
      bulma_tag("Dark", color = "dark"),
      bulma_tag("Light", color = "light"),
      bulma_tag("White", color = "white"),
      bulma_tag("Primary", color = "primary"),
      bulma_tag("Link", color = "link"),
      bulma_tag("Info", color = "info"),
      bulma_tag("Success", color = "success"),
      bulma_tag("Warning", color = "warning"),
      bulma_tag("Danger", color = "danger")
    ),
    bulma_block(
      bulma_title("Light Colors"),
      bulma_tag("Primary", color = "primary", light = TRUE),
      bulma_tag("Link", color = "link", light = TRUE),
      bulma_tag("Info", color = "info", light = TRUE),
      bulma_tag("Success", color = "success", light = TRUE),
      bulma_tag("Warning", color = "warning", light = TRUE),
      bulma_tag("Danger", color = "danger", light = TRUE)
    ),
    bulma_block(
      bulma_title("Different Sizes"),
      bulma_tag("Normal [Default]", size = "normal", color = "link"),
      bulma_tag("Medium", size = "medium", color = "primary"),
      bulma_tag("Large", size = "large", color = "info"),
      bulma_tags(
        size = "medium",
        bulma_tag("All"),
        bulma_tag("Medium"),
        bulma_tag("Size")
      ),
      bulma_tags(
        size = "large",
        bulma_tag("All"),
        bulma_tag("Large"),
        bulma_tag("Size")
      ),
      bulma_tags(
        size = "medium",
        bulma_tag("Medium"),
        bulma_tag("Normal", size = "normal"),
        bulma_tag("Medium"),
        bulma_tag("Large", size = "large"),
        bulma_tag("Medium")
      ),
      bulma_block(
        bulma_title("Modifiers"),
        bulma_tag("Rounded", style = "rounded"),
        bulma_tag(style = "delete", container = htmltools::a)
      ),
      bulma_block(
        bulma_title("Combinations"),
        bulma_tag("Bar", color = "success", add_delete = TRUE),
        bulma_tag("Hello", color = "warning", size = "medium", add_delete = TRUE),
        bulma_tag("World", color = "danger", size = "large", add_delete = TRUE,
                  delete_size = NULL)
      ),
      bulma_block(
        bulma_title("List of Tags"),
        bulma_tags(
          lapply(numbers[1:3], bulma_tag)
        ),
        bulma_tags(
          lapply(numbers, bulma_tag)
        )
      ),
      bulma_block(
        bulma_title("Tag Addons"),
        bulma_tags(
          bulma_tag("Package"),
          bulma_tag("shinyBulma", color = "primary"),
          addons = TRUE
        ),
        bulma_tags(
          bulma_tag("Alex Smith", color = "danger"),
          bulma_tag(color = "dark", style = "delete", container = htmltools::a),
          addons = TRUE
        )
      )
    )
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/elements/tag/}{Bulma Tag}
}
