% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{download_bulma_themes}
\alias{download_bulma_themes}
\title{Download Pre-Compiled Bulma Themes}
\usage{
download_bulma_themes(themes = "all", gh_user = NULL, gh_password = NULL)
}
\arguments{
\item{themes}{[\code{character(n)}: \sQuote{all}]\cr the names of the themes to be
installed. If \sQuote{all}, all available themes are downloaded.}

\item{gh_user, gh_password}{[\code{character(1)}: \sQuote{NULL}]\cr GitHub credentials,
see \emph{Note}.}
}
\value{
list (invisible), the used download links for the fonts and the themes
respectively.
}
\description{
This function downloads pre-compiled CSS files and Google font files for the
\href{https://jenil.github.io/bulmaswatch/}{bulmaswatch themes}. These files are
hosted on this package's \href{https://github.com/thothal/shinyBulma2}{Github repository}.
}
\section{Note}{

To dynamically get the URLs from the GitHub repo, we make use of
\href{https://docs.github.com/en/rest}{GitHub's API}. Without authentification, one can make
up to 60 calls to the API per hour. In very rare occasions, you may exceed this limit
(for instance if you are calling this function several times). The limit for
authenticated API calls is 1000 calls. Thus, you can provide your credentials to
this function to significantly increase the limit.
}

\examples{
\dontrun{
download_bulma_themes("flatly")
}
}
