% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_bulma_themes}
\alias{get_bulma_themes}
\title{Get List of Installed Bulma Themes}
\usage{
get_bulma_themes()
}
\value{
character vector, all installed bulma themes.
}
\description{
This function looks into the \verb{lib/} folder located in the package's base folder and
returns all found themes in the corresponding subfolder. Themes are compiled from
bulmaswatch.
}
\examples{
get_bulma_themes()
}
\seealso{
\url{https://jenil.github.io/bulmaswatch/}
}
