% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bulma_color}
\alias{validate_bulma_color}
\title{Validate Proper Bulma Color Definition}
\usage{
validate_bulma_color(x, context = c("text", "background"), must_be_key = FALSE)
}
\arguments{
\item{x}{[\code{character(n)}]\cr
The color to validate.}

\item{context}{[\code{character(1)}: \sQuote{text}]\cr
If \dQuote{text} the color must work as a text color via \code{has-text-{color}}. If
\dQuote{background} the color must work as a background color via
\code{has-background-{color}}.}

\item{must_be_key}{[\code{logical(1)}: \sQuote{FALSE}]\cr
If \sQuote{TRUE}, the color to checked must be a key of the SASS \verb{$colors} map.
These colors can be used by some elements directly via \code{is-{color}} instead of
\code{has-text-{color}}.}
}
\value{
The proper bulma color class, if possible. Otherwise an error is raised.
}
\description{
\code{validate_bulma_color} checks that the argument is a valid bulma color and can be used
by the \verb{has-(text|background)-.*} helpers respectively.
}
\details{
This function uses the bulma config data sets \link{bulma_config} and
\link{bulmaswatch_config} respectively to check whether the argument is valid. The relevant
column in these datasets are scraped from the official bulma documentation.
}
\section{Note}{
In order to determine whether the bulma or the bulmaswatch config dataset should be
used, this function relies on an internal variable which points to the theme used in
\code{bulma_page} (if any was specified.). Thus, the results may differ depending on the
last call to \code{bulma_page}. At the time being, the same colors are defined for the
default \code{bulma} theme and the \code{bulmaswatch} themes. However, this may change in the
future. Thus, we opted for this approach.
}

\examples{
validate_bulma_color("black")
validate_bulma_color("primary-light")
## standard bulma config will be used
validate_bulma_color("link")
if ("flatly" \%in\% get_bulma_themes()) {
   ## flatly theme config will be used
   bulma_page(theme = "flatly")
   validate_bulma_color("link")
   ## reset theme
   bulma_page()
}
validate_bulma_color("success", "background")
}
