% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{validate_bulma_column_size}
\alias{validate_bulma_column_size}
\alias{validate_bulma_column_offset}
\title{Validate Proper Width or Offset Definition for a Bulma Column}
\usage{
validate_bulma_column_size(x)

validate_bulma_column_offset(x)
}
\arguments{
\item{x}{[\code{character(1)} or \code{integer(1)}]\cr
The size or offset to validate.}
}
\value{
The proper bulma size or offset class, if possible. Otherwise an error.
}
\description{
Checks that the argument is a valid column width or offset specifier. Valid values are
integers between 1 and 12, or strings like \code{two-fifth} or \code{full}. For the size \code{narrow}
is also a valid option. Additionally, to the original specifiers
(cf. \href{https://bulma.io/documentation/columns/sizes/}{column sizes}), you can also
abbreviate the number-ish strings, e.g. \code{"2/5"} instead of \code{"two-fifth"}.
}
\examples{
validate_bulma_column_size(1)
validate_bulma_column_size("one-third")
validate_bulma_column_size("1/1")
validate_bulma_column_size("narrow-widescreen")

validate_bulma_column_offset(6)
validate_bulma_column_offset("two-thirds")
validate_bulma_column_offset("4/5")

\dontrun{
## these will each raise an error
validate_bulma_column_size(13)
validate_bulma_column_size(1:2)
validate_bulma_column_size(11.3)
validate_bulma_column_size("one-sixth")
validate_bulma_column_size("one.third")

validate_bulma_column_offset("narrow")
validate_bulma_column_offset("full")
validate_bulma_column_offset("1/1")
}
}
