% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bulma_size}
\alias{validate_bulma_size}
\title{Validate Proper Bulma Size Definition}
\usage{
validate_bulma_size(
  x,
  normal_to_null = TRUE,
  prefix = NULL,
  disallow = NULL,
  additional = NULL
)
}
\arguments{
\item{x}{[\code{character(n)}]\cr
The size to validate.}

\item{normal_to_null}{[\code{logical(1)}: \sQuote{TRUE}]\cr
If \sQuote{TRUE} size \sQuote{normal} is replaced by \sQuote{NULL}. This is
useful if \sQuote{normal} is anyways the default.}

\item{prefix}{[\code{character(1)}: \sQuote{NULL}]\cr
The prefix used for the size class, should be \sQuote{is} (or equivalently
\sQuote{NULL}) or \sQuote{are}.}

\item{disallow}{[\code{character(n)}: \sQuote{NULL}]\cr
Disallow certain classes in the validation. This is needed in case an element
does not support all of the standard bulma sizes (e.g. \verb{<tags>}).}

\item{additional}{[\code{character(n)}: \sQuote{NULL}]\cr
Include additional classes in the validation. This is needed for elements like
\code{bulma_hero} which allow sizes like \code{fullheight}.}
}
\value{
The proper bulma size class, if possible. Otherwise an error is raised.
}
\description{
Validate Proper Bulma Size Definition
}
\details{
This function simply compares its arguments with the list of valid bulma
sizes and returns the proper class or throws an error if an invalid size is passed.
Valid sizes are:
\itemize{
\item \code{small}
\item \code{normal}
\item \code{medium}
\item \code{large}
\item plus additional elements from \code{additional}
}
}
\examples{
validate_bulma_size(c("small", "normal", "medium", "large"), FALSE)
validate_bulma_size("normal") ## returns NULL as it is replaced
## in this form this is pretty useless, but in a function it makes
## snese to allow additional elements to be valid
validate_bulma_size("fullheight", additional = "fullheight")

\dontrun{
## these will all raise an error
validate_bulma_size("xlarge")
validate_bulma_size("small", disallow = "small")
}
}
