% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bulma_unit}
\alias{validate_bulma_unit}
\alias{validate_bulma_size}
\alias{validate_bulma_offset}
\alias{validate_bulma_gap}
\alias{get_bulma_media_breakpoints}
\title{Validate Proper Width, Offset or Gap Definition for a Bulma Column Layout}
\usage{
validate_bulma_unit(x, type = c("size", "offset", "gap"))

validate_bulma_size(x)

validate_bulma_offset(x)

validate_bulma_gap(x)

get_bulma_media_breakpoints(x)
}
\arguments{
\item{x}{[\code{character(n)} or \code{integer(n)}]\cr
The sizes, offsets or gaps to validate.}

\item{type}{[\code{character(1)}: \sQuote{size}]\cr
The type of unit we want to validate. Must be one of \code{size}, \code{offset} or \code{gap}.}
}
\value{
The proper bulma size, offset or gap classes, if possible. Otherwise an error
is raised.
\code{get_bulma_media_breakpoints} returns a character vector with all valid breakpoints.
}
\description{
\code{validate_bulma_unit} checks that the argument is a valid column
width, offset or gap specifier, respectively.
}
\details{
\verb{validate_bulma_\{size|offset|gap\}} are convenience wrappers which set \code{type} to the
respective value.

\code{get_bulma_media_breakpoints} returns all valid media breakpoints.

Valid values for sizes and offsets are integers between 1 and 12, or strings like
\code{two-fifth} or \code{full}. For the size \code{narrow} is also a valid option. Additionally, to
the original specifiers
(cf. \href{https://bulma.io/documentation/columns/sizes/}{column sizes}), you can also
abbreviate the number-ish strings, e.g. \code{"2/5"} instead of \code{"two-fifth"}.

Gap values are integers between 0 and 8 or the string \dQuote{gapless}.

All values can additionally be suffixed by valid media breakpoint qualifiers
in which case these sizes are only valid on screens which meet at least the size given
by the breakpoints.

Multiple values can be passed. This makes sense, if we want to define different values
for different media breakpoints. Warnings are issued, either if
duplicated media breakpoints are detected, or if we provide multiple values without
media breakpoints.
}
\examples{
validate_bulma_size(1)
validate_bulma_size("one-third")
validate_bulma_size("1/1")
validate_bulma_size("narrow-tablet")
validate_bulma_size(c("1-touch", "2-desktop", "3-fullhd"))

validate_bulma_offset(6)
validate_bulma_offset("two-thirds")
validate_bulma_offset("4/5-desktop")
validate_bulma_offset(c("two-thirds-touch", "11-desktop", "4/5-fullhd"))

validate_bulma_gap("gapless")
validate_bulma_gap(1)
validate_bulma_gap("2-touch")
validate_bulma_gap(c("1-touch", "2-mobile", "3-desktop"))

get_bulma_media_breakpoints()

\dontrun{
## these will each raise an error
validate_bulma_size(13)
validate_bulma_size(11.3)
validate_bulma_size("one-sixth")
validate_bulma_size("one.third")
validate_bulma_size("1-touchscreen")
validate_bulma_size(c("1-fullhd", "2"))
validate_bulma_size(c("1-fullhds", "2-fullhds"))

validate_bulma_offset("narrow")
validate_bulma_offset("full")
validate_bulma_offset("1/1")
validate_bulma_offset("1/3-4k")

validate_bulma_gap(9)
validate_bulma_gap("one-third")

## these will raise warnings
validate_bulma_size(1:3)
validate_bulma_offset(c("1-fullhd", "2-fullhd"))
}
}
