% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{Bulma-Elements}
\alias{Bulma-Elements}
\alias{bulma_block}
\alias{bulma_box}
\alias{bulma_content}
\title{Create Various bulma Elements}
\usage{
bulma_block(...)

bulma_box(...)

bulma_content(..., size = NULL)
}
\arguments{
\item{...}{[\verb{html tags} or \verb{html attributes}]\cr
Elements to include within the container.}

\item{size}{[\code{character(1)}: \sQuote{NULL}]\cr
Text size for the \code{content} container.}
}
\value{
The bulma element container.
}
\description{
These functions create the respective bulma containers.
}
\examples{

if (interactive() && requireNamespace("shiny", quietly = TRUE)) {

  blind_text <- tagList(h1("Hello World"),
                        p("Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
                           "Nulla accumsan, metus ultrices eleifend gravida, nulla",
                           "nunc varius lectus, nec rutrum justo nibh eu lectus. Ut",
                           "vulputate semper dui. Fusce erat odio, sollicitudin vel",
                           "erat vel, interdum mattis neque."),
                        h2("Second Level"),
                        p("Curabitur accumsan turpis pharetra",
                          strong("augue tincidunt"), "blandit. Quisque",
                          "condimentum maximus mi, sit amet commodo arcu rutrum id.",
                          "Proin pretium urna vel cursus venenatis.",
                          "Suspendisse potenti. Etiam mattis sem rhoncus lacus",
                          "dapibus facilisis. Donec at dignissim dui. Ut et neque",
                          "nisl."),
                        tags$ul(tags$li("In fermentum leo eu lectus mollis, quis",
                                        "dictum mi aliquet."),
                                tags$li("Morbi eu nulla lobortis, lobortis est in,",
                                        "fringilla felis."),
                                tags$li("Aliquam nec felis in sapien venenatis",
                                        "viverra fermentum nec lectus."),
                                tags$li("Ut non enim metus.")))


  ui <- bulma_page(
    h1("Block", class = "title is-1"),
    bulma_block("This text is within a", strong("block"), "."),
    bulma_block("This text is within a", strong("second block"), ". Lorem ipsum",
                "dolor sit amet, consectetur adipiscing elit. Aenean efficitur sit",
                "amet massa fringilla egestas. Nullam condimentum luctus turpis."),
    bulma_block("This text is within a ", strong("third block"), ". This block",
                "has no margin at the bottom."),
    h1("Box", class = "title is-1"),
    bulma_box("I'm a box"),
    h1("Content", class = "title is-1"),
    h2("Normal Size", class = "title is-3"),
    bulma_content(blind_text),
    h2("Small Size", class = "title is-3"),
    bulma_content(blind_text, size = "small")
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
\seealso{
\href{https://bulma.io/documentation/elements/}{Bulma Elements}
}
