% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulma_config}
\alias{bulma_config}
\title{Colors and Font Used in Bulma}
\format{
A data frame with 85 rows and 6
variables:
\describe{
\item{theme}{[\code{character(1)}], the name of the theme (here this equals always
\sQuote{bulma} and is included to be inline with
\link{bulmaswatch_config}.)}
\item{group}{[\code{character(1)}], either \sQuote{color} or \sQuote{font} describing
whether the variable describes a color or the font.}
\item{variable}{[\code{character(1)}], the name of color or \sQuote{name} or
\sQuote{id} for fonts describing the name or the id.}
\item{value}{[\code{character(1)}], the hex color code or the name of the font used.}
\item{is_color_map_key}{[\code{logical(1)}], is the particular color a key in the SASS
\verb{$colors} map? These colors are used in \code{is-{color}} classes, e.g. \code{is-primary} or
\code{is-black}.}
\item{color_class}{[\code{character(1)}], either \code{NA} if the variable is not a color or
there is no color helper class to use this color. Otherwise it contains a comma
separated string with elements \sQuote{text} if it can be used via
\code{has-text-{color}} and \sQuote{background} it it can be used via
\code{has-background-{color}}.}
}
}
\source{
\url{https://bulma.io/documentation/customize/variables/#derived-variables}
}
\usage{
bulma_config
}
\description{
A dataset containing all the colors and the imported Google font in the bulma
stylesheet.
}
\seealso{
\link{bulmaswatch_config}
}
\keyword{datasets}
