% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bulma_size}
\alias{validate_bulma_size}
\title{Validate Proper Bulma Size Definition}
\usage{
validate_bulma_size(x, normal_to_null = TRUE, prefix = NULL)
}
\arguments{
\item{x}{[\code{character(n)}]\cr
The size to validate.}

\item{normal_to_null}{[\code{logical(1)}: \sQuote{TRUE}]\cr
If \sQuote{TRUE} size \sQuote{normal} is replaced by \sQuote{NULL}. This is
useful if \sQuote{normal} is anyways the default.}

\item{prefix}{[\code{character(1)}: \sQuote{NULL}]\cr
The prefix used for the size class, should be \sQuote{is} (or equivalently
\sQuote{NULL}) or \sQuote{are}.}
}
\value{
The proper bulma size class, if possible. Otherwise an error is raised.
}
\description{
Validate Proper Bulma Size Definition
}
\details{
This function simply compares its arguents with the list of valid bulma
sizes and returns the proper class or throws an error if an invalid size is passed.
Valid sizes are:
\itemize{
\item \code{small}
\item \code{normal}
\item \code{medium}
\item \code{large}
}
}
\examples{
validate_bulma_size(c("small", "normal", "medium", "large"), FALSE)
validate_bulma_size("normal") ## returns NULL as it is replaced

\dontrun{
## this will raise an error
validate_bulma_size("xlarge")
}
}
